+++ appdaemon/plugins/hass/hassapi.py
@@ -375,8 +375,9 @@ class Hass(adbase.ADBase, adapi.ADAPI):
         """
         namespace = self._get_namespace(**kwargs)
         await self._check_entity(namespace, entity_id)
+        kwargs["entity_id"] = entity_id
 
-        await self.get_entity_api(namespace, entity_id).turn_on(**kwargs)
+        await self.call_service("homeassistant/turn_on", **kwargs)
 
     @utils.sync_wrapper
     @hass_check
@@ -410,13 +411,13 @@ class Hass(adbase.ADBase, adapi.ADAPI):
             >>> self.turn_off("scene.bedroom_on")
 
         """
-        namespace = self._get_namespace(**kwargs)
         domain, _ = await self.split_entity(entity_id)
+        kwargs["entity_id"] = entity_id
 
         if domain == "scene":
-            await self.get_entity_api(namespace, entity_id).turn_on(**kwargs)
+            await self.call_service("homeassistant/turn_on", **kwargs)
         else:
-            await self.get_entity_api(namespace, entity_id).turn_off(**kwargs)
+            await self.call_service("homeassistant/turn_off", **kwargs)
 
     @utils.sync_wrapper
     @hass_check
@@ -447,8 +448,9 @@ class Hass(adbase.ADBase, adapi.ADAPI):
         """
         namespace = self._get_namespace(**kwargs)
         await self._check_entity(namespace, entity_id)
+        kwargs["entity_id"] = entity_id
 
-        await self.get_entity_api(namespace, entity_id).toggle(**kwargs)
+        await self.call_service("homeassistant/toggle", **kwargs)
 
     @utils.sync_wrapper
     @hass_check